rasterizer_cache: Separate into common directory
