+++ src/video_core/rasterizer_cache/surface_params.cpp
@@ -3,8 +3,8 @@
 // Refer to the license.txt file included.
 
 #include "common/alignment.h"
-#include "video_core/renderer_opengl/gl_rasterizer_cache.h"
-#include "video_core/renderer_opengl/gl_surface_params.h"
+#include "video_core/rasterizer_cache/rasterizer_cache.h"
+#include "video_core/rasterizer_cache/surface_params.h"
 
 namespace OpenGL {
