+++ src/video_core/renderer_opengl/texture_filters/texture_filter_base.h
@@ -6,7 +6,7 @@
 
 #include "common/common_types.h"
 #include "common/math_util.h"
-#include "video_core/renderer_opengl/gl_surface_params.h"
+#include "video_core/rasterizer_cache/surface_params.h"
 
 namespace OpenGL {
